requirements.txt
flask
werkzeug
twilio
groq
vocode
elevenlabs
deepgram-sdk
python-dotenv
pydantic
requests
uuid